+++ backend_admin/src/main/java/com/example/backend_admin/elastic/model/entity/ProductDetailDocument.java
@@ -1,12 +1,14 @@
-package com.example.backend_admin.elastic.entity;
+package com.example.backend_admin.elastic.model.entity;
 
+import lombok.Getter;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.elasticsearch.annotations.Document;
 import org.springframework.data.elasticsearch.annotations.Field;
 import org.springframework.data.elasticsearch.annotations.FieldType;
 
-@Document(indexName = "productsearch")
-public class ProductSearchDocument {
+@Getter
+@Document(indexName = "productdetail")
+public class ProductDetailDocument {
     @Id
     private String id;
 
@@ -16,6 +18,9 @@ public class ProductSearchDocument {
     @Field(type = FieldType.Integer)
     private int customerIdx;
 
-    @Field(type = FieldType.Text)
-    private String keyword;
+    @Field(type = FieldType.Integer)
+    private int productIdx;
+
+    @Field(type = FieldType.Integer)
+    private int category;
 }
