+++ backend_admin/src/main/java/com/example/backend_admin/elastic/repository/OrdersDocumentRepository.java
@@ -1,7 +1,6 @@
 package com.example.backend_admin.elastic.repository;
 
-import com.example.backend_admin.elastic.entity.LoginDocument;
-import com.example.backend_admin.elastic.entity.OrdersDocument;
+import com.example.backend_admin.elastic.model.entity.OrdersDocument;
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
 
 import java.util.Date;
