+++ backend_admin/src/main/java/com/example/backend_admin/elastic/repository/ProductSearchDocumentRepository.java
@@ -1,6 +1,6 @@
 package com.example.backend_admin.elastic.repository;
 
-import com.example.backend_admin.elastic.entity.ProductSearchDocument;
+import com.example.backend_admin.elastic.model.entity.ProductSearchDocument;
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
 
 public interface ProductSearchDocumentRepository extends ElasticsearchRepository<ProductSearchDocument, String> {
